Move all engine move functions to engine_wrapper
